fix: preserve business logic of account going into overdraft

+++ Converters/AccountConverter.cs
@@ -20,7 +20,7 @@ namespace MiniBank.Converters
                 var account = new AccountFactory().Create(type);
 
                 account.ID = dataReader.GetString(dataReader.GetOrdinal("ID"));
-                account.Balance = dataReader.GetFloat(dataReader.GetOrdinal("Balance"));
+                account.Balance = (float) dataReader.GetDouble(dataReader.GetOrdinal("Balance"));
                 account.OwnerID = dataReader.GetString(dataReader.GetOrdinal("OwnerID"));
 
                 return account;
